# api.yaml
openapi: 3.0.3
info:
  title: GoChat IM API
  description: |
    A comprehensive instant messaging API built with Go, providing real-time chat functionality,
    user management, friend relationships, group chats, and file uploads.

    ## Features
    - JWT-based authentication
    - Real-time messaging via WebSocket
    - User registration and profile management
    - Friend relationship management
    - Group chat functionality
    - File upload with deduplication
    - Online status tracking

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## WebSocket Connection
    Real-time messaging uses WebSocket at `/ws?token=<jwt-token>`

  version: "1.0.0"
  contact:
    name: GoChat API Support
    email: support@gochat.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.gochat.com/api/v1
    description: Production server

# Security schemes
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "BAD_REQUEST"
        message:
          type: string
          description: Human readable error message
          example: "Invalid request data"
        details:
          type: string
          description: Additional error details
          example: "Field 'email' is required"
      required:
        - code
        - message

    # Success response schema
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          description: Success code (always 0)
          example: 0
        message:
          type: string
          description: Success message
          example: "success"
        data:
          type: object
          description: Response data

    # User model
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: User unique identifier
          example: 1
        phone:
          type: string
          description: User phone number (11 digits)
          pattern: '^\d{11}$'
          example: "13800138000"
        nickname:
          type: string
          description: User display name
          minLength: 2
          maxLength: 50
          example: "John Doe"
        avatar:
          type: string
          description: Avatar image URL
          example: "avatars/user123.png"
        gender:
          type: integer
          description: User gender (0=unset, 1=male, 2=female)
          enum: [0, 1, 2]
          example: 1
        signature:
          type: string
          description: Personal signature
          maxLength: 200
          example: "Hello, I'm using GoChat!"
        created_at:
          type: string
          format: date-time
          description: Account creation time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last profile update time
          example: "2023-06-01T12:30:00Z"
      required:
        - id
        - phone
        - nickname

    # Message model
...
paths:
...
  /group/{id}/members:
    get:
      summary: Get group members
      description: Retrieve list of group members
      operationId: getGroupMembers
      tags:
        - Group Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Group members retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          members:
                            type: array
                            items:
                              allOf:
                                - $ref: '#/components/schemas/User'
                                - type: object
                                  properties:
                                    joined_at:
                                      type: string
                                      format: date-time
                                      description: When they joined the group
                                      example: "2023-01-01T00:00:00Z"
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Add group members
      description: Add new members to a group (owner only)
      operationId: addGroupMembers
      tags:
        - Group Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                member_ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                  description: User IDs to add to the group
                  example: [5, 6]
              required:
                - member_ids
            examples:
              add_members:
                summary: Add members request
                value:
                  member_ids: [5, 6]
      responses:
        '200':
          description: Members added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid member IDs or users already in group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group owner can add members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Conversation endpoints
...
security:
  - bearerAuth: []